# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EcoTrack API",
        "description": "Système de traçabilité intelligente pour la Supply Chain (Mobile & IoT).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8003",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/api/allproducts": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Lister les produits (paginé)",
                "description": "Retourne la liste paginée des produits avec leur société (company).",
                "operationId": "ef825e12ed18442cd413ec4488e7d344",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page (défaut 20)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des produits",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products": {
            "post": {
                "tags": [
                    "Produits"
                ],
                "summary": "Créer un nouveau produit",
                "description": "Détail d'un produit par UUID.\n\nGET /api/products/{uuid}",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "sku"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sac de Café Arabica 25kg"
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "CAF-AR-001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{uuid}/qrcode": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Générer l'image du QR Code",
                "operationId": "20b140e10ee77252caa33b27abaf1915",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image du QR Code (SVG)"
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    }
                }
            }
        },
        "/api/products/{uuid}/scan": {
            "post": {
                "tags": [
                    "Traçabilité"
                ],
                "summary": "Enregistrer un scan officiel (Mobile ou IoT)",
                "description": "Cette route est utilisée par les terminaux de terrain pour mettre à jour le statut et la position GPS.",
                "operationId": "93354ecf2b98d91f1ac5295d96b1a7ec",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "processed",
                                            "in_transit",
                                            "delivered"
                                        ]
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 6.36
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.43
                                    },
                                    "location_name": {
                                        "type": "string",
                                        "example": "Entrepôt Cotonou"
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "example": {
                                            "temp": 4.5,
                                            "humidity": 20
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Scan enregistré avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{uuid}/history": {
            "get": {
                "tags": [
                    "Consultation Publique"
                ],
                "summary": "Afficher la timeline du produit (Scan QR Code)",
                "description": "Route publique accessible par le consommateur final pour vérifier la provenance et l'historique du produit.",
                "operationId": "97ef2caf37235a32dc895e1de99a1c1a",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historique récupéré"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre token reçu lors du login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}